+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+### v1.1.0 (12/09/2021)
+* Added new custom minimal MinGW64 windres resource compiler
+* Added new Icon and Assembly Data options using the new resource compiler
+* Increased key complexity to avoid general key scans
+* Fixed general small bugs
+* Optimized code
 ### v1.0.0 (09/09/2021)
 * Initial release